[MCUX-83760] Fix MISRA C-2012 Rule 10.3
Implicit conversion of length * element_size from essential type signed 32-bit int to different or narrower essential type unsigned 32-bit int.
Cast variables to uint32_t and arithm operation to ULONG.

+++ lib/rpmsg_lite/porting/environment/rpmsg_env_threadx.c
@@ -658,7 +658,7 @@ int32_t env_create_queue(void **queue, int32_t length, int32_t element_size)
     }
 
     if (TX_SUCCESS == tx_queue_create((TX_QUEUE *)queue_ptr, NULL, ((uint32_t)element_size / RL_WORD_SIZE),
-                                      (VOID *)msgq_buffer_ptr, (length * element_size)))
+                                      (VOID *)msgq_buffer_ptr, (ULONG)((uint32_t)length * (uint32_t)element_size)))
     {
         *queue = (void *)queue_ptr;
         return 0;
